% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_snps.R
\name{are_snps}
\alias{are_snps}
\title{Determine if the one or more strings are a SNP in HGVS sequence.
This is the vectorized version of \link{is_snp}}
\usage{
are_snps(s)
}
\arguments{
\item{s}{string in HVGS format to be parsed,
for example, \code{NP_000585.2:p.Gly144Asp}}
}
\description{
Determine if the one or more strings are a SNP in HGVS sequence.
This is the vectorized version of \link{is_snp}
}
\examples{
# These are SNPs
are_snps("NP_001771.1:p.Ser91Pro")
are_snps("NP_001254627.1:p.Val237Met")
are_snps("NP_055635.3:p.Ala55Gly")

# A synonymous mutation at the DNA level is no SNP
are_snps(s = "NP_001771.1:p.Phe89=")

# A deletion is no SNP
are_snps("NP_001771.1:p.Leu64_Gly74del")

# A frame shift at the DNA level is no SNP
are_snps("NP_001771.1:p.Asp167fs")
}
