\name{Collection-class}
\alias{Collection-class}
\title{Collection-class}
\usage{
  new_Collection(...)
}
\description{
  A container of S4 objects with a shared environment that
  may contain metadata.
}
\seealso{
  \code{\link{makeCollectionClass}},
  \code{\link{collectionConstructor}},
  \code{\link{collectionValidator}},
  \code{\link{collectionShower}}.rt
}
\keyword{classes}
\keyword{internal}

