\name{collectionShower}
\alias{collectionShower}
\title{Create a show-method for a Collection class.}
\usage{
  collectionShower(showFun, numOfElements = 6,
    linesPerElement = NULL)
}
\arguments{
  \item{showFun}{show function for Collection class
  elements.}

  \item{numOfElements}{Number of elements to show as head
  and tail.}

  \item{linesPerElement}{How many lines per element before
  the show-string is elipsized.}
}
\description{
  Create a show-method for a Collection class.
}
\keyword{internal}

