\name{collectionValidator}
\alias{collectionValidator}
\title{Create a validity-checking method for a Collection class.}
\usage{
  collectionValidator(Class)
}
\arguments{
  \item{Class}{Name of the collection class.}
}
\description{
  Create a validity-checking method for a Collection class.
}
\keyword{internal}

