% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_msa.R
\name{create_msa}
\alias{create_msa}
\title{Create a multiple sequence alignment using COBALT}
\usage{
create_msa(
  fasta_text,
  cobalt_options = cobaltr::create_cobalt_options(),
  cobalt_folder = cobaltr::get_default_cobalt_folder()
)
}
\arguments{
\item{fasta_text}{text of a FASTA file}

\item{cobalt_options}{a set of COBALT options,
as can be created by \link[cobaltr]{create_cobalt_options}}

\item{cobalt_folder}{the folder where COBALT is installed in.
Use \link[cobaltr]{get_default_cobalt_folder} to get the default COBALT folder}
}
\value{
the FASTA text of the MSA
}
\description{
Create a multiple sequence alignment using COBALT
}
