% Built by roxygen3 0.1
\docType{class}
\name{doi-class}
\alias{doi-class}
\title{doi-class}
\description{
A Digital Object Identifier.
}
\section{Slots}{


\describe{
   \item{\code{doi}:}{A character vector.}
}
}

\section{Generics with methods for doi}{
\itemize{
  \item \code{\link{[}}
  \item \code{\link{browsePubmed}}
  \item \code{\link{c}}
  \item \code{\link{show}}
}

}

\keyword{classes}
\keyword{internal}
