% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_protein_ids_from_variations.R
\name{extract_protein_ids_from_variations}
\alias{extract_protein_ids_from_variations}
\title{Extract the protein IDs from variations}
\usage{
extract_protein_ids_from_variations(variations)
}
\description{
Extract the protein IDs from variations
}
