% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences_from_protein_ids.R
\name{fetch_sequences_from_protein_ids}
\alias{fetch_sequences_from_protein_ids}
\title{Fetch the sequence from one or more protein ID}
\usage{
fetch_sequences_from_protein_ids(protein_ids)
}
\value{
a character vector with protein sequences,
one per protein ID
}
\description{
Fetch the protein sequence using one or more protein IDs
}
