% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences_from_subject_ids.R
\name{fetch_sequences_from_subject_ids}
\alias{fetch_sequences_from_subject_ids}
\title{Fetch the sequence from a protein ID}
\usage{
fetch_sequences_from_subject_ids(subject_ids, verbose = FALSE)
}
\arguments{
\item{subject_ids}{one of more IDs of a subject}

\item{verbose}{set to TRUE for more output}
}
\value{
text that can be saved as a FASTA file
}
\description{
Fetch the protein sequence using a protein ID
}
