% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_human_membrane_protein_gene_ids.R
\name{get_all_human_membrane_protein_gene_ids}
\alias{get_all_human_membrane_protein_gene_ids}
\title{Get all the human membrane protein gene names}
\usage{
get_all_human_membrane_protein_gene_ids()
}
\description{
Get all the human membrane protein gene names
}
