% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_name_snp_variation_topology.R
\name{get_gene_name_snp_variation_topology}
\alias{get_gene_name_snp_variation_topology}
\title{From a gene name, get the SNP ID, variation and topology}
\usage{
get_gene_name_snp_variation_topology(gene_name)
}
\description{
From a gene name, get the SNP ID, variation and topology
}
