% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_ids_from_gene_name.R
\name{get_snp_ids_from_gene_name}
\alias{get_snp_ids_from_gene_name}
\title{Get all the SNP IDs for a certain gene}
\usage{
get_snp_ids_from_gene_name(gene_name)
}
\description{
Get all the SNP IDs for a certain gene
}
