% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_variations_in_protein.R
\name{get_snp_variations_in_protein}
\alias{get_snp_variations_in_protein}
\title{From a SNP id, get the variations at the protein level.
Returns \code{character(0)} if there are none}
\usage{
get_snp_variations_in_protein(snp_id)
}
\description{
From a SNP id, get the variations at the protein level.
Returns \code{character(0)} if there are none
}
