% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_variations_in_protein_from_snp_id.R
\name{get_snp_variations_in_protein_from_snp_id}
\alias{get_snp_variations_in_protein_from_snp_id}
\title{From a SNP id, get the variations at the protein level.
Returns \code{character(0)} if there are none}
\usage{
get_snp_variations_in_protein_from_snp_id(snp_id)
}
\arguments{
\item{snp_id}{the ID of a SNP}
}
\description{
From a SNP id, get the variations at the protein level.
Returns \code{character(0)} if there are none
}
