\name{makeCollectionClass}
\alias{makeCollectionClass}
\title{Create a Collection Class Definition}
\usage{
  makeCollectionClass(elementClass)
}
\arguments{
  \item{elementClass}{Class name of the elements.}
}
\value{
  A constructor function for
  \code{<elementClass>List}-Class.
}
\description{
  Create a subclass to \code{\linkS4class{Collection}}
  named \code{<elementClass>List} and return a constructor
  function.
}
\details{
  The constructor function is generated by a call to
  \code{\link{collectionConstructor}}, a simple
  validity-checking method is supplied by a call to
  \code{\link{collectionConstructor}}.

  If a \code{show} method has been defined for the
  Collection elements, a generic \code{show} method can be
  constructed by a call to \code{\link{collectionShower}}.
}
\seealso{
  \code{\linkS4class{Collection}},
  \code{\link{collectionConstructor}},
  \code{\link{collectionValidator}},
  \code{\link{collectionShower}}.
}
\keyword{internal}

