% Built by roxygen3 0.1
\docType{package}
\name{ncbi}
\alias{ncbi}
\alias{ncbi-package}
\title{Retrieve data from NCBI.}
\description{
Retrieve data from NCBI.
}
\section{description}{


  Important functions:
  \itemize{
   
    \item \code{\link{pubmed}}: Retrieve records from the \emph{PubMed}
    database.
    \item \code{\link{protein}}: Retrieve records from the \emph{protein}
    database.
    \item \code{\link{nucleotide}}: Retrieve records from the
    \emph{nuccore} database.
    \item \code{\link{GSS}}: Retrieve records from the
    \emph{nucgss} database.   
    \item \code{\link{EST}}: Retrieve records from the
    \emph{nucest} database.        
    \item \code{\link{taxon}}: Retrieve records from the
    \emph{Taxonomy} database.
    \item \code{\link{taxonDB}}: Retrieve records from a local install of
    the \emph{Taxonomy} database.    
    
  }
}

\author{
Gerhard Schöfl \email{gschofl@yahoo.de}
}
