% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hgvs.R
\name{parse_hgvs}
\alias{parse_hgvs}
\title{Parse a HGVS variation description}
\usage{
parse_hgvs(s)
}
\arguments{
\item{s}{string in HVGS format to be parsed,
for example, \code{NP_000585.2:p.Gly144Asp}}
}
\value{
the variation as a list
}
\description{
Parse a HGVS variation description
}
\note{
see \url{http://varnomen.hgvs.org/} for the HGVS format
}
