% Built by roxygen3 0.1
\name{parseGenBank}
\alias{parseGenBank}
\title{Parse GenBank flatfiles}
\usage{
  parseGenBank(gb)
}
\arguments{
  \item{gb}{An \linkS4class{efetch} instance or a character vector
containing a GenBank flatfile.}
}

\value{
A \linkS4class{gbRecord} instance.
}
\description{
Parse GenBank flatfiles
}
