% Built by roxygen3 0.1
\name{author}
\alias{abstract}
\alias{author}
\alias{browsePubmed}
\alias{cites}
\alias{doi}
\alias{issue}
\alias{journal}
\alias{pmid}
\alias{title}
\title{PubMed accessors

Access the various slots of \code{\link[utils]{bibentry}} instances.}
\usage{
  author(x, ...)

  title(x, ...)

  abstract(x, ...)

  journal(x, ...)

  issue(x, ...)

  pmid(x, ...)

  doi(x, ...)

  cites(x, ...)

  browsePubmed(x, ...)
}
\arguments{
  \item{x}{A \code{\link[utils]{bibentry}} instance.}

  \item{...}{Further arguments passed on to methods.}
}

\section{Method signatures for generic author}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic title}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic abstract}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic journal}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic issue}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic pmid}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic doi}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic cites}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic browsePubmed}{
\itemize{
  \item \linkS4class{doi}
  \item \linkS4class{pubmed}
}

}

\keyword{methods}
