\docType{methods}
\name{getAuthor}
\alias{[}
\alias{[,doi-method}
\alias{[,pubmed-method}
\alias{browsePubmed}
\alias{browsePubmed,doi-method}
\alias{browsePubmed,pubmed-method}
\alias{c}
\alias{c,doi-method}
\alias{c,pubmed-method}
\alias{getAbstract}
\alias{getAbstract,pubmed-method}
\alias{getAuthor}
\alias{getAuthor,pubmed-method}
\alias{getCites}
\alias{getCites,pubmed-method}
\alias{getDoi}
\alias{getDoi,pubmed-method}
\alias{getIssue}
\alias{getIssue,pubmed-method}
\alias{getJournal}
\alias{getJournal,pubmed-method}
\alias{getPmid}
\alias{getPmid,pubmed-method}
\alias{getTitle}
\alias{getTitle,pubmed-method}
\alias{show}
\alias{show,doi-method}
\alias{show,pubmed-method}
\title{Access pubmed objects}
\arguments{
  \item{x}{A \code{\linkS4class{pubmed}} object.}

  \item{...}{Further arguments passed on to methods.}
}
\description{
  Access the various slots of \code{\linkS4class{pubmed}}
  objects.
}

