% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_human_membrane_proteins.R
\name{search_for_human_membrane_proteins}
\alias{search_for_human_membrane_proteins}
\title{Get the IDs of human membrane proteins.}
\usage{
search_for_human_membrane_proteins(verbose = FALSE)
}
\description{
Get the IDs of human membrane proteins.
}
\seealso{
Search terms are documented at
\url{https://www.ncbi.nlm.nih.gov/books/NBK25499/#_chapter4_ESearch_}.
The Protein Advanced Search Builder at
\url{https://www.ncbi.nlm.nih.gov/protein/advanced}
helps to generate a specific search term
}
