% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_protein_ids.R
\name{search_protein_ids}
\alias{search_protein_ids}
\title{Search for a protein its ID}
\usage{
search_protein_ids(accession, verbose = FALSE)
}
\arguments{
\item{accession}{protein accession (?), for example, \code{4ZW9}.}
}
\description{
Search for the ID of a human protein.
}
\seealso{
Search terms are documented at
\url{https://www.ncbi.nlm.nih.gov/books/NBK25499/#_chapter4_ESearch_}.
The Protein Advanced Search Builder at
\url{https://www.ncbi.nlm.nih.gov/protein/advanced}
helps to generate a specific search term
}
