% Built by roxygen3 0.1
\docType{class}
\name{taxonList-class}
\alias{taxonList-class}
\title{taxonList}
\description{
An extension of \link{list} that holds only \linkS4class{taxon}
instances.
}
\section{Class hierarchy for taxonList}{
Sub classes:
\itemize{
  \item \linkS4class{Lineage}
}
Super classes:
\itemize{
  \item \linkS4class{list}
  \item \linkS4class{vector}
  \item \linkS4class{listOrDataFrame}
  \item \linkS4class{vectorORfactor}
}

}

\section{Generics with methods for taxonList}{
\itemize{
  \item \code{\link{sciName}}
  \item \code{\link{show}} (1 methods defined for subclasses)
  \item \code{\link{taxId}}
  \item \code{\link{taxRank}}
}

}

\keyword{classes}
