% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_protein_ids_from_variations.R
\name{extract_protein_ids_from_variations}
\alias{extract_protein_ids_from_variations}
\title{Extract the protein IDs from variations}
\usage{
extract_protein_ids_from_variations(variations)
}
\arguments{
\item{variations}{one or more variations.
For example, \code{NP_009089.4:p.Val723Gly} is a variation.}
}
\description{
Extract the protein IDs from variations
}
