% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_hgvs_in_tmh.R
\name{is_hgvs_in_tmh}
\alias{is_hgvs_in_tmh}
\title{Determine if a HGVS is a mutation in a TMH part of membrane protein}
\usage{
is_hgvs_in_tmh(hgvs)
}
\description{
Determine if a HGVS is a mutation in a TMH part of membrane protein
}
