% Built by roxygen3 0.1
\docType{package}
\name{ncbi}
\alias{ncbi}
\alias{ncbi-package}
\title{Full-text search of NCBI's databases}
\description{
Full-text search of NCBI's databases
}
\section{description}{


  Important functions:
  \itemize{
   
    \item \code{pubmed}: Search the \emph{PubMed} database.
    
    \item \code{protein}: Search the \emph{Protein} database.

    \item \code{nucleotide}: Search the \emph{Nucleotide} database.
    
    \item \code{\link{taxon}}: Search the \emph{Taxonomy} database.
    
  }
}

\author{
Gerhard Schöfl \email{gschofl@yahoo.de}
}
