% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hgvs.R
\name{parse_hgvs}
\alias{parse_hgvs}
\title{Parse a HGVS variation description}
\usage{
parse_hgvs(s)
}
\arguments{
\item{s}{string in HVGS format to be parsed}
}
\value{
the variation as a list
}
\description{
Parse a HGVS variation description
}
