\name{parseGenBank}
\alias{parseGenBank}
\title{Parse GenBank flatfiles}
\usage{
  parseGenBank(gb)
}
\arguments{
  \item{gb}{An \linkS4class{efetch} instance, a character
  vector containing a GenBank flatfile, or a file path to a
  GenBank file.}
}
\value{
  A \linkS4class{gbRecord} instance.
}
\description{
  Parse GenBank flatfiles
}

