% Built by roxygen3 0.1
\name{getAuthor}
\alias{browsePubmed}
\alias{getAbstract}
\alias{getAuthor}
\alias{getCites}
\alias{getDoi}
\alias{getIssue}
\alias{getJournal}
\alias{getPmid}
\alias{getTitle}
\title{PubMed accessors

Access the various slots of \code{\link[utils]{bibentry}} instances.}
\usage{
  getAuthor(x, ...)

  getTitle(x, ...)

  getAbstract(x, ...)

  getJournal(x, ...)

  getIssue(x, ...)

  getPmid(x, ...)

  getDoi(x, ...)

  getCites(x, ...)

  browsePubmed(x, ...)
}
\arguments{
  \item{x}{A \code{\link[utils]{bibentry}} instance.}

  \item{...}{Further arguments passed on to methods.}
}

\section{Method signatures for generic getAuthor}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic getTitle}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic getAbstract}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic getJournal}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic getIssue}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic getPmid}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic getDoi}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic getCites}{
\itemize{
  \item \linkS4class{pubmed}
}

}

\section{Method signatures for generic browsePubmed}{
\itemize{
  \item \linkS4class{doi}
  \item \linkS4class{pubmed}
}

}

\keyword{methods}
