\docType{methods}
\name{getTaxID}
\alias{getAuthority}
\alias{getLineage}
\alias{getOtherName}
\alias{getParentTaxID}
\alias{getRank}
\alias{getScientificName}
\alias{getTaxID}
\title{Taxon accessors}
\arguments{
  \item{x}{A \linkS4class{Taxon} or
  \linkS4class{TaxonList}.}

  \item{...}{Further arguments passed on to methods.}
}
\description{
  Access the slots of \linkS4class{Taxon} or
  \linkS4class{TaxonList} objects.
}

