% Built by roxygen3 0.1
\name{taxId}
\alias{authority}
\alias{lineage}
\alias{parent}
\alias{sciName}
\alias{synonym}
\alias{taxId}
\alias{taxRank}
\title{Taxon accessors}
\usage{
  taxId(x, ...)

  sciName(x, ...)

  taxRank(x, ...)

  synonym(x, ...)

  authority(x, ...)

  lineage(x, ...)

  parent(x, ...)
}
\arguments{
  \item{x}{A \linkS4class{taxon} or \linkS4class{taxonList}.}

  \item{...}{Further arguments passed on to methods.}
}

\description{
Access the various slots of \linkS4class{taxon} or \linkS4class{taxonList}
instances.
}
\section{Method signatures for generic taxId}{
\itemize{
  \item \linkS4class{taxon}
  \item \linkS4class{taxonList}
}

}

\section{Method signatures for generic sciName}{
\itemize{
  \item \linkS4class{taxon}
  \item \linkS4class{taxonList}
}

}

\section{Method signatures for generic taxRank}{
\itemize{
  \item \linkS4class{taxon}
  \item \linkS4class{taxonList}
}

}

\section{Method signatures for generic synonym}{
\itemize{
  \item \linkS4class{taxon}
}

}

\section{Method signatures for generic authority}{
\itemize{
  \item \linkS4class{taxon}
}

}

\section{Method signatures for generic lineage}{
\itemize{
  \item \linkS4class{taxon}
}

}

\section{Method signatures for generic parent}{
\itemize{
  \item \linkS4class{taxon}
}

}

\keyword{methods}
